feat: add avatar to list view
Used the prebuilt helper that should have all accessability solutions.
If a use has not uploaded an image might be best to use a letter to
indicate who the child is. For accounts with only one child its better
to remove the picture but that can be a global update (single child
mode)